+++ compiler/rustc_query_impl/Cargo.toml
@@ -16,5 +16,6 @@ rustc_middle = { path = "../rustc_middle" }
 rustc_query_system = { path = "../rustc_query_system" }
 rustc_serialize = { path = "../rustc_serialize" }
 rustc_span = { path = "../rustc_span" }
+rustc_thread_pool = { path = "../rustc_thread_pool" }
 tracing = "0.1"
 # tidy-alphabetical-end
