+++ compiler/rustc_query_impl/src/lib.rs
@@ -23,10 +23,11 @@ use rustc_middle::query::values::Value;
 use rustc_middle::ty::TyCtxt;
 use rustc_query_system::dep_graph::SerializedDepNodeIndex;
 use rustc_query_system::query::{
-    CycleError, CycleErrorHandling, QueryCache, QueryMap, QueryMode, QueryState,
+    CycleError, CycleErrorHandling, QueryCache, QueryMode, QueryState,
 };
 use rustc_span::{ErrorGuaranteed, Span};
 
+pub use crate::job::{QueryMap, break_query_cycles, print_query_stack};
 pub use crate::plumbing::{QueryCtxt, query_key_hash_verify_all};
 use crate::plumbing::{encode_all_query_results, try_mark_green};
 use crate::profiling_support::QueryKeyStringCache;
@@ -34,6 +35,7 @@ pub use crate::profiling_support::alloc_self_profile_query_strings;
 
 mod error;
 mod execution;
+mod job;
 #[macro_use]
 mod plumbing;
 mod profiling_support;
